XRDDEV-429 improve configurability
